Move state and test into own files
